1.update opencv to 3.4.16 2.update darknet arrch64 options

+++ modules/webcam/snaps.cpp
@@ -89,7 +89,7 @@ void csnaps::work(){
 		newframe.append(buf, nrecv);
 		nsize -= nrecv;
 	}
-	printf(" recv:%d\n", newframe.size());
+	printf(" recv:%d\n", (int)newframe.size());
 	setframe(&newframe[0], (int)newframe.size());
 	return;
 }
